+++ CHANGELOG.md
@@ -34,6 +34,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `fetch_bikes` to `datasets`
 - `simulate_qa` to `stream`
 - `online_qa_score` to `model_selection`
+- `Pipeline`'s now have a `debug_one(x)` method
 
 ### Modified